+++ CHANGELOG.markdown
@@ -5,7 +5,8 @@ Version 0.7.6 Beta
 ------------------
 
 - Throw exceptions when browsercap ini file not loaded but specified
-- Inform user via php error log when browsercap ini file has been updated and needs loading 
+- Inform user via php error log when browsercap ini file has been updated and needs loading
+- BrowserCap ini file now updated whenever update available on server
 
 Version 0.7.5 Beta
 ------------------